# Fix performance degradation