--- a/model/modeldecoder/error_test.go
+++ b/model/modeldecoder/error_test.go
@@ -246,7 +246,7 @@ func TestErrorEventDecode(t *testing.T) {
 				Config:      test.cfg,
 			}, batch)
 			require.NoError(t, err)
-			assert.Equal(t, test.e, &batch.Errors[0])
+			assert.Equal(t, test.e, batch.Errors[0])
 		})
 	}
 }