--- a/model/modeldecoder/transaction_test.go
+++ b/model/modeldecoder/transaction_test.go
@@ -362,7 +362,7 @@ func TestTransactionEventDecode(t *testing.T) {
 				Config:      test.cfg,
 			}, batch)
 			require.NoError(t, err)
-			assert.Equal(t, test.e, &batch.Transactions[0])
+			assert.Equal(t, test.e, batch.Transactions[0])
 		})
 	}
 }