## Motivation/summary

Change the model.Batch type to store slices of
pointers, rather than structs. This enables us
to reuse the model.Batch (with the new Reset
method), and also avoids some memory copying
in modeldecoder (i.e. where it was decoding to
a heap allocated object, and then copying to
slice element when dereferencing).

In the future we might want to revert to storing
slices of structs, but only when we can safely
reuse a batch. i.e. pool model.Batch objects,
recycling once the publisher has transformed the
model objects to beat.Events.

benchstat, comparing to master:

```
name                                          old time/op    new time/op    delta
BackendProcessor/heavy.ndjson/NoRateLimit-12    31.4ms ± 2%    28.9ms ± 4%   -7.98%  (p=0.008 n=5+5)

name                                          old speed      new speed      delta
BackendProcessor/heavy.ndjson/NoRateLimit-12  12.7MB/s ± 2%  13.8MB/s ± 4%   +8.69%  (p=0.008 n=5+5)

name                                          old alloc/op   new alloc/op   delta
BackendProcessor/heavy.ndjson/NoRateLimit-12    9.94MB ± 0%    6.33MB ± 0%  -36.30%  (p=0.002 n=6+6)

name                                          old allocs/op  new allocs/op  delta
BackendProcessor/heavy.ndjson/NoRateLimit-12      115k ± 0%      113k ± 0%   -1.83%  (p=0.002 n=6+6)
```

## Checklist

- [x] I have signed the [Contributor License Agreement](https://www.elastic.co/contributor-agreement/).
- [x] My code follows the style guidelines of this project (run `make check-full` for static code checks and linting)
- [x] I have rebased my changes on top of the latest master branch
~- [ ] I have made corresponding changes to the documentation~
~- [ ] I have added tests that prove my fix is effective or that my feature works~
- [x] New and existing [**unit** tests](https://github.com/elastic/apm-server/blob/master/TESTING.md) pass locally with my changes
~- [ ] I have updated [CHANGELOG.asciidoc](https://github.com/elastic/apm-server/blob/master/CHANGELOG.asciidoc)~

## How to test these changes

`go test -count=6 -benchtime=10s -benchmem -bench=BenchmarkBackendProcessor/heavy.ndjson/NoRateLimit -run=NONE github.com/elastic/apm-server/processor/stream`

(Run with master and this branch, use benchstat to compare)

## Related issues

Closes #3754